moved sdk and created ge_sdk package
